rebase

+++ common/xbps-src/shutils/cross.sh
@@ -1,12 +1,12 @@
 # vim: set ts=4 sw=4 et:
 
 remove_pkg_cross_deps() {
-    local rval= tmplogf=
+    local rval= tmplogf= prevs=0
     [ -z "$XBPS_CROSS_BUILD" ] && return 0
 
     cd $XBPS_MASTERDIR || return 1
     msg_normal "${pkgver:-xbps-src}: removing autocrossdeps, please wait...\n"
-    tmplogf=$(mktemp)
+    tmplogf=$(mktemp) || exit 1
 
     if [ -z "$XBPS_REMOVE_XCMD" ]; then
         source_file $XBPS_CROSSPFDIR/${XBPS_CROSS_BUILD}.sh
@@ -14,7 +14,18 @@ remove_pkg_cross_deps() {
     fi
 
     $XBPS_REMOVE_XCMD -Ryo > $tmplogf 2>&1
-    if [ $? -ne 0 ]; then
+    rval=$?
+    while [ $rval -eq 0 ]; do
+        local curs=$(stat -c %s $tmplogf)
+        if [ $curs -eq $prevs ]; then
+            break
+        fi
+        prevs=$curs
+        $XBPS_REMOVE_XCMD -Ryo >> $tmplogf 2>&1
+        rval=$?
+    done
+
+    if [ $rval -ne 0 ]; then
         msg_red "${pkgver:-xbps-src}: failed to remove autocrossdeps:\n"
         cat $tmplogf && rm -f $tmplogf
         msg_error "${pkgver:-xbps-src}: cannot continue!\n"
@@ -24,23 +35,24 @@ remove_pkg_cross_deps() {
 
 prepare_cross_sysroot() {
     local cross="$1"
+    local statefile="$XBPS_MASTERDIR/.xbps-${cross}-done"
 
-    [ -z "$cross" -o "$cross" = "" ] && return 0
+    [ -z "$cross" -o "$cross" = "" -o -f $statefile ] && return 0
 
     # Check for cross-vpkg-dummy available for the target arch, otherwise build it.
-    pkg_available cross-vpkg-dummy $cross
+    pkg_available 'cross-vpkg-dummy>=0.33_1' $cross
     if [ $? -eq 0 ]; then
         $XBPS_LIBEXECDIR/build.sh cross-vpkg-dummy cross-vpkg-dummy pkg $cross init || return $?
     fi
 
-    check_installed_pkg cross-vpkg-dummy-0.17_1 $cross
+    check_installed_pkg cross-vpkg-dummy-0.30_1 $cross
     [ $? -eq 0 ] && return 0
 
     msg_normal "Installing $cross cross pkg: cross-vpkg-dummy ...\n"
-    errlog=$(mktemp)
+    errlog=$(mktemp) || exit 1
     $XBPS_INSTALL_XCMD -Syfd cross-vpkg-dummy &>$errlog
     rval=$?
-    if [ $rval -ne 0 -a $rval -ne 17 ]; then
+    if [ $rval -ne 0 ]; then
         msg_red "failed to install cross-vpkg-dummy (error $rval)\n"
         cat $errlog
         rm -f $errlog
@@ -48,10 +60,12 @@ prepare_cross_sysroot() {
     fi
     rm -f $errlog
     # Create top level symlinks in sysroot.
-    XBPS_ARCH=$XBPS_TARGET_MACHINE xbps-reconfigure -r $XBPS_CROSS_BASE -f base-directories base-files &>/dev/null
+    XBPS_ARCH=$XBPS_TARGET_MACHINE xbps-reconfigure -r $XBPS_CROSS_BASE -f base-files &>/dev/null
     # Create a sysroot/include and sysroot/lib symlink just in case.
     ln -s usr/include ${XBPS_CROSS_BASE}/include
     ln -s usr/lib ${XBPS_CROSS_BASE}/lib
+
+    touch -f $statefile
 
     return 0
 }
@@ -65,13 +79,13 @@ install_cross_pkg() {
     pkg_available cross-${XBPS_CROSS_TRIPLET}
     rval=$?
     if [ $rval -eq 0 ]; then
-        $XBPS_LIBEXECDIR/build.sh cross-${XBPS_CROSS_TRIPLET} cross-${XBPS_CROSS_TRIPLET} pkg || return $rval
+        $XBPS_LIBEXECDIR/build.sh cross-${XBPS_CROSS_TRIPLET} cross-${XBPS_CROSS_TRIPLET} pkg || return $?
     fi
 
     check_installed_pkg cross-${XBPS_CROSS_TRIPLET}-0.1_1
     [ $? -eq 0 ] && return 0
 
-    errlog=$(mktemp)
+    errlog=$(mktemp) || exit 1
     msg_normal "Installing $cross cross compiler: cross-${XBPS_CROSS_TRIPLET} ...\n"
     $XBPS_INSTALL_CMD -Syfd cross-${XBPS_CROSS_TRIPLET} &>$errlog
     rval=$?
